Added the fetch request to the background process
The fetch request will fetch the requested resource (via url). The
response is cached and served up on subsequent requests provided that the
force argument is not given in the request and that cache is not older
than the request's age argument. If no age is given, it defaults to one
hour.