Added recently_visited script to enhance the Recently Visited list
Unlike the previous implementation, this one is much simpler since it just
uses the list that OkCupid provides rather than trying to track your
history itself. However, it'll still keep track of more than just the most
recent 6 and it's still scrollable. Plus, the clear button is back and
will clear BetterCupid's log as well as OkCupid's.

This implementation is not compatible with the previous one.