Major major rewrite.
Stripped out a lot of dev crap in favor of using built-in browser tools.
Stripped out a lot of other crap in favor of doing things better.
Moved from using "scripts" to using "modules"
 - modules are Markup+CSS+Script
 - modules have their own db functionality for getting/setting
    their data and settings
 - modules have an "options" varient for the Options page
The background page can now better handle upgrading versions without
 loss of data
The options page can now better support module options.
A better system was implemented for handling "default" settings

A lot of functionality has been lost in this version, but recent changes
to OkCupid broke most of it anyhow and the app has been unusable, so this
is the first step in getting it back to being usable again.