Added Module.get(moduleID)
Once a module is loaded, it'll be available via Module.get(). No immediate
need for this but it's something I've been thinking about. I'll need to
flesh it out more once I have a use-case for it so that dependencies
can be set to ensure that a dependant module is available when the
depending module is loaded.