Various Module improvements, including module dependencies
Added module.enabled([val]) as a shorthand for module.db.get("enabled")
and module.db.set("enabled", val)

Added module.state() which returns a string indicating the state of
the module. States include: ready (ready to be loaded), waiting (waiting
for dependencies to load), failed (cannot be loaded or failed to load),
and loaded (loaded successfully).

The module loader will now delay the loading of a module until all of its
dependencies have been loaded. This functionality works recursively. If
a dependency is disabled or fails to load, the module loader will not try
to load the dependant module. Circular references in module dependencies
are not checked and will throw the module loader into an infinite loop.