Starting to add background page support.
It's been awhile since I've worked on this, but it looks like I was
prepping to include background page support for querying notifications
like new messages and quiver updates. Not sure why I didn't commit small
pieces at a time as I went, but I'll try my best to summarize everything.

Created the DB lib that can interact with localstorage or any other kind
of object to store and retrive data in JSON format. DB objects will support
event delegation so that things can be informed of when a module's DB is
updated and reflect the changes.

Created the Pigeon lib to assist in Tab interactions.

Various changes to support the background page.

Some bug fixes.

New themes module. This may not last. There are other dedicated extensions
for applying themes to any page.