More manifest v2 updates
The regular module pages still work just fine with the inline js,
but the option pages don't work that way under the new security rules.
So I had to break their JS into separate files and create a new function
for getting the module object. It's a bit hacky, but it'll work for now.
This project is due for some major restructuring.