modules.json enhancements and added Module.inject and Module.get
modules.json is now much more explicit, and nothing is assumed about
where module files are placed or what files there are.
In addition to defining the module's name, id, and description, the file
acts like a manifest and defines the defaults (removing the need for a
defaults.json), the script files for injecting into the site, the option
files for the options page, and the background files for the background
page.

The new Module.inject method takes a string (scripts, options, or
background) and injects the relevant resources into the current context.
It will inject css, html, and js files defined in the modules.json for
all modules and create the module objects in the process.
The second argument is a callback that's executed once all modules are
fully loaded. The callback takes a modules array which consists of all
loaded modules. Modules that are disabled are not loaded in "scripts" mode.

Module.get takes a module id and returns the module object.

module.path now treats relative paths starting with a "/" character as
being relative to the extension's root directory. Paths that don't start
with this character are relative to the module's directory (which is
still assumed to be the /modules/module-name path).

core.rootPath will now return a relative path if loaded into an extension
page (background or options) and an absolute path that includes the
chrome-extension directive when loaded in a content script.

core.assert was added and used in a number of places.

load and load-options were updated to use Module.inject for the heavy
lifting. They should ultimately be replaced.

The module scripts were updated to use Module.get rather than the
globally defined module object (which I don't think would work with the
new way of doing things anyhow).

The new Module class design is still not perfect. It assumes that inject
will be the first thing used and it assumes that inject will never be
called twice on the same page.. even with different parameters. Doing so
would be very bad since the page module containers and option module
containers use the same ID. Still, it's getting better.