Gulp updates - file downloads, nunjuck filters
Added a function so that gulp can download files.. was originally going
to download the changelog and about like the app currently does, but
only on a production build, and then keep it in the dist/options.html so
it doesn't need to load each time -- this proved troublesome. The app
currently uses jquery to navigate the dom structure .. I'm not sure of a
node equivalent for that, plus the code is a bit ridiculous anyhow.
Instead, I converted the changelog to a json format that I can feed into
options.html template with the manifest, and I could potentially use
this to generate the markdown for the wiki changelog as well.

To facilitate the above, I need to check on major.minor version numbers,
not rev.. so I typically use parseFloat. That doesn't work in the
template. I could use regex, but that didn't work either, so I created a
custom filter called 'version' that uses regex. This even works on
assignments.. so it works pretty well for tracking that so I know when
to create a new version section.

I also fixed a bug with the gulp-options task.. it took in the done arg,
but never called it, so the task would only run once.

I also noticed I had commented out all the individual build watches ..
they're uncommented now.