Begin support for testing and docs
For module authors to come in, better cupid will need some docs to help
them get started, and will need real tests to make sure things don't
break (I break things alot) -- start with unit tests for now, work on
integration tests later.

Mocha will provide the unit test framework in a browser within the
extension's context so that it has access to the chrome extension APIs
to test everything properly.

Chai will provide the assertions to write the tests.

Also adding gulp-preprocess so that during dev some functions can run
extra checks to make sure they're being used properly, but in production
they'll skip them for efficiency.

The gulpfile has undergone a lot of changes to add in the doc and
testing stuff. I added the optional opener lib for opening the browser
to the docs when built since that's how I usually do it with yuidoc, but
I wasn't happy with yuidoc so currently using jsdoc. Was also using
opener to open a custom extension I was working on (based on
chrome-extensions-reloader) that handled reloading chrome extensions and
their windows.. but the experience was very jarring. Currently using
LiveJS instead, which will not be added to the build for prod.