Changelog notes and refresh support
The gulpfile was using require() to get the json files, but apparently
that caches them in addition to the cache I had in place. Added a 1
second cache timeout and now using JSON.parse(getFile()) to read them in
so that changes to them will be reflected when LiveJS refreshes.